Added page_file_format to page
Page dataclass now has a new field for page_file_format that does not
have to be specified. This field corresponds to the format the text is
in (confluencewiki, markdown, html).

If the format is not specified, there is a helper function that looks at
the file extension and tries to guess it.

Progress on #35.